requirements.txt
numpy==1.11.1
Pillow==3.3.0
PyYAML==3.11
requests==2.10.0
wordcloud==1.2.1
